Spilt the API endpoint from main to routers for easier readability
